// src/store/login/types.ts
export enum Types {
  SAVE_AUTH_LOGIN = 'SAVE_AUTH_LOGIN',
  SAVE_FIREBASE_TOKEN = 'SAVE_FIREBASE_TOKEN',
  SAVE_USER_INFO = 'SAVE_USER_INFO',
  SAVE_USER_INFO_CORE = 'SAVE_USER_INFO_CORE',
  SHOW_LOADING = 'SHOW_LOADING',
  HIDE_LOADING = 'HIDE_LOADING',
  USER_INTERACT_WEB = 'USER_INTERACT_WEB',
  RESET = 'RESET'
}